Evals/RubberDuck/requirements.txt
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0
requests>=2.31.0 